User Roles, UserAuthorities updated and default Users & UserInfos added
UserService.java updated to create a UserInfo when a user registers AND when an administrator adds a new user from the admin panel. UserService.java updated to delete a UserInfo associated with a User before the User entity is discarded.
Changed password minlength to "4" in thelearningzone\src\main\webapp\app\account\register\register.component.html